changes
git-svn-id: svn+ssh://howdy.physics.nyu.edu/usr/local/svn/kcorrect/trunk@203 75dff457-8751-0410-ab1a-cd701654ce6b

+++ src/k_brent.c
@@ -8,12 +8,12 @@
 #define SHFT(a,b,c,d) (a)=(b);(b)=(c);(c)=(d);
 #define SIGN(a,b) ((b) >= 0.0 ? fabs(a) : -fabs(a))
 
-double k_brent(double ax, double bx, double cx, double (*f)(double), 
-							 double tol, double *xmin)
+float k_brent(float ax, float bx, float cx, float (*f)(float), 
+							float tol, float *xmin)
 {
 	int iter;
-	double a,b,d,etemp,fu,fv,fw,fx,p,q,r,tol1,tol2,u,v,w,x,xm;
-	double e=0.0;
+	float a,b,d,etemp,fu,fv,fw,fx,p,q,r,tol1,tol2,u,v,w,x,xm;
+	float e=0.0;
 
 	a=(ax < cx ? ax : cx);
 	b=(ax > cx ? ax : cx);
