+++ src/k_interpolate.c
@@ -14,13 +14,13 @@
  */
 
 /* linear interpolate in a given vector; out of bounds --> 0 */
-double k_interpolate(double currpos,
-										 double vals[],
-										 double pos[],
+float k_interpolate(float currpos,
+										 float vals[],
+										 float pos[],
 										 IDL_LONG n)
 {
 	unsigned long i,ip1;
-	double sp,currval;
+	float sp,currval;
 
 	k_locate(pos, (unsigned long) n, currpos, &i);
 	if(i>=n-1) return(vals[n-1]);
@@ -33,15 +33,15 @@ double k_interpolate(double currpos,
 
 /* linear interpolate in a given vector; out of bounds --> 0;
  * assume x-axis is given equally spaced */
-double k_interpolate_es(double currpos,
-											double vals[],
-											double pos[],
+float k_interpolate_es(float currpos,
+											float vals[],
+											float pos[],
 											IDL_LONG n)
 {
 	long i,ip1;
-	double sp,currval;
+	float sp,currval;
 
-	i=(long) floor((double) n*(currpos-pos[0])/
+	i=(long) floor((float) n*(currpos-pos[0])/
 													(2*pos[n-1]-pos[n-2]-pos[0]));
 	if(i>=n) return(vals[n-1]);
 	if(i<0) return(vals[0]);
