+++ src/k_utils.c
@@ -4,12 +4,12 @@
 #define NR_END 1
 #define FREE_ARG char*
 
-double *k_vector(long nl, long nh)
-/* allocate a double vector with subscript range v[nl..nh] */
+float *k_vector(long nl, long nh)
+/* allocate a float vector with subscript range v[nl..nh] */
 {
-	double *v;
+	float *v;
 
-	v=(double *)malloc((size_t) ((nh-nl+1+NR_END)*sizeof(double)));
+	v=(float *)malloc((size_t) ((nh-nl+1+NR_END)*sizeof(float)));
 	if (!v) {
 		fprintf(stderr,"allocation failure in vector()");
 		exit(1);
@@ -17,8 +17,8 @@ double *k_vector(long nl, long nh)
 	return v-nl+NR_END;
 }
 
-void k_free_vector(double *v, long nl, long nh)
-/* free a double vector allocated with vector() */
+void k_free_vector(float *v, long nl, long nh)
+/* free a float vector allocated with vector() */
 {
 	free((FREE_ARG) (v+nl-NR_END));
 }
